refactor(headless-styles): separate out icon button tests
